# ensembl/datacheck/swagger.yml
openapi: "3.0.2"

info:
  title: "Datachecks"
  version: "1.0"

tags:
  - name: Retrieve datacheck information, and submit and monitor datacheck jobs

paths:
  /datacheck/names:
    get:
      tags:
        - Retrieve datacheck information
      summary: "Return all datachecks"
      responses:
        200:
          description: Details of all datachecks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/names"

  /datacheck/names/{name}:
    get:
      tags:
        - Retrieve datacheck information
      summary: "Return a specified datacheck"
      parameters:
        - $ref: "#/components/parameters/name"
      responses:
        200:
          description: Details of a specified datacheck.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/names"

  /datacheck/groups:
    get:
      tags:
        - Retrieve datacheck information
      summary: "Return all datachecks, categorised by group"
      responses:
        200:
          description: Details of all datachecks, categorised by group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/groups"

  /datacheck/groups/{group}:
    get:
      tags:
        - Retrieve datacheck information
      summary: "Return datachecks in a specified group"
      parameters:
        - $ref: "#/components/parameters/group"
      responses:
        200:
          description: Details of all datachecks in the specified group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/groups"

  /datacheck/types:
    get:
      tags:
        - Retrieve datacheck information
      summary: "Return all datachecks, categorised by type (critical or advisory)"
      responses:
        200:
          description: Details of all datachecks, categorised by type (critical or advisory).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/types"

  /datacheck/types/{type}:
    get:
      tags:
        - Retrieve datacheck information
      summary: "Return datachecks for the specified type"
      parameters:
        - $ref: "#/components/parameters/type"
      responses:
        200:
          description: Details of all datachecks for the specified type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/types"

  /datacheck/search/{keyword}:
    get:
      tags:
        - Retrieve datacheck information
      summary: "Return datachecks whose name or description partially match the keyword"
      parameters:
        - $ref: "#/components/parameters/keyword"
      responses:
        200:
          description: Details of all datachecks whose name or description partially match the keyword (case insensitive).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/keywords"

components:
  parameters:
    name:
      name: name
      in: path
      description: A datacheck name
      required: true
      schema:
        type: string
        example: ForeignKeys

    group:
      name: group
      in: path
      description: A datacheck group
      required: true
      schema:
        type: string
        example: corelike

    type:
      name: type
      in: path
      description: A datacheck type
      required: true
      schema:
        type: string
        enum: ["critical", "advisory"]
        example: critical

    keyword:
      name: keyword
      in: path
      required: true
      schema:
        type: string
        example: xref

  schemas:
    datacheck:
      title: datacheck
      type: object
      properties:
        name:
          type: string
          description: Name of the datacheck.
          example: "SchemaPatchesApplied"

        package_name:
          type: string
          description: Name of the datacheck Perl module.
          example: "Bio::EnsEMBL::DataCheck::Checks::SchemaPatchesApplied"

        description:
          type: string
          description: Description of the datacheck.
          example: "Check that schema patches are up-to-date"

        datacheck_type:
          type: string
          description: Type of the datacheck, either critical or advisory.
          example: "critical"

        groups:
          type: array
          items:
            type: string
            description: Group to which the datacheck belongs.
            example: "corelike"
          description: List of groups to which the datacheck belongs.
          example: ["core", "corelike"]

    names:
      title: names
      type: object
      properties:
        name:
          type: array
          items:
            $ref: "#/components/schemas/datacheck"

    groups:
      title: groups
      type: object
      properties:
        group:
          type: array
          items:
            $ref: "#/components/schemas/datacheck"

    types:
      title: types
      type: object
      properties:
        type:
          type: array
          items:
            $ref: "#/components/schemas/datacheck"

    keywords:
      title: keyword
      type: object
      properties:
        keyword:
          type: array
          items:
            $ref: "#/components/schemas/datacheck"